major update

+++ src/main/java/fast_fix/repository/FuelStationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FuelStationRepository extends JpaRepository<FuelStation, Integer> {
+public interface FuelStationRepository extends JpaRepository<FuelStation, Long> {
 }
